+++ R/boilerplate-theta.R
@@ -156,5 +156,17 @@ ts_auto_theta <- function(.data, .date_col, .value_col, .rsamp_obj,
         )
     )
 
+    # Add attributes
+    attr(output, ".tune") <- .tune
+    attr(output, ".grid_size") <- .grid_size
+    attr(output, ".cv_assess") <- .cv_assess
+    attr(output, ".cv_skip") <- .cv_skip
+    attr(output, ".cv_slice_limit") <- .cv_slice_limit
+    attr(output, ".best_metric") <- .best_metric
+    attr(output, ".bootstrap_final") <- .bootstrap_final
+    attr(output, ".mode") <- "regression"
+    attr(output, ".parsnip_engine") <- "theta"
+    attr(output, ".function_family") <- "boilerplate"
+
     return(invisible(output))
 }